fix possible  'uchar' overflow and unexpected results' overwriting
in theory, saving non-first result should be harmless and valid,

+++ profanity.cl
@@ -678,7 +678,7 @@ __kernel void profanity_iterate(__global mp_number * const pDeltaX, __global mp_
 
 void profanity_result_update(const size_t id, __global const uchar * const hash, __global result * const pResult, const uchar score, const uchar scoreMax) {
 	if (score && score > scoreMax) {
-		uchar hasResult = atomic_inc(&pResult[score].found); // NOTE: If "too many" results are found it'll wrap around to 0 again and overwrite last result. Only relevant if global worksize exceeds MAX(uint).
+		uint hasResult = atomic_inc(&pResult[score].found); // NOTE: If "too many" results are found it'll wrap around to 0 again and overwrite last result. Only relevant if global worksize exceeds MAX(uint).
 
 		// Save only one result for each score, the first.
 		if (hasResult == 0) {
